Add Major changes in typography and Recent Posts
This is my first iteration on the minimo theme. The fonts have been
replaced with my new fav, Darker Grotesque and the old timer Crimson
Text.

Some font size adjustments and related bugs have been fixed. The newer
theme is cleaner and spacious. The post titles are given prority and the
fonts are uniform and spaced out. The overall size of the website is
slightly increased to help with readability.

And finally there is no need for custom.css in the main blog folder. Now
the styling and the content have been completely separated in two
distint repos.

Also, I am maintaining this theme for my main use only. If someone wants
to use a similar style, he is rather encouraged to use minimo the OG
theme this one is based on.

Peace! Yo xypnox