Add animated border and typography improvements
With this commit I have introduced animated borders on hover for several
links and navigation elements:

 - header links
 - tags in the footer
 - Taxonomy cloud tags

This has been achieved through mixin named border-animate that takes in
the size of the border in px.

Also, the color of icons has been changed to match the accent color. The
icons in the footer for social links have been left as is.

The typography for next and previous post and accompanying icons have
been tweaked.

The sample site's configuration has been tweaked to use the new theme
name.

---

xypnox